[chore] #74 mapper 분리
