+++ app/src/main/java/org/sopt/pingle/presentation/mapper/PinMapper.kt
@@ -33,17 +33,3 @@ fun PinEntity.toMarkerModel(): MarkerModel {
 
     return markerModel
 }
-
-fun PingleEntity.isCompleted() = maxParticipants == curParticipants
-
-fun PingleEntity.convertToCalenderDetail(): String {
-    val localDate = LocalDate.parse(date, DateTimeFormatter.ISO_LOCAL_DATE)
-    val startTime = LocalTime.parse(startAt, DateTimeFormatter.ISO_LOCAL_TIME)
-    val endTime = LocalTime.parse(endAt, DateTimeFormatter.ISO_LOCAL_TIME)
-
-    return buildString {
-        append("${localDate.year}년 ${localDate.monthValue}월 ${localDate.dayOfMonth}일\n")
-        append("${startTime.format(DateTimeFormatter.ofPattern("HH:mm"))} ~ ")
-        append("${endTime.format(DateTimeFormatter.ofPattern("HH:mm"))}")
-    }
-}
